IndexedDB: Handle generated keys up to 2^53 https://bugs.webkit.org/show_bug.cgi?id=85114
Source/WebCore:

The spec defines the behavior for generated keys up to 2^53
(the maximum integer storable as an ECMAScript number) and
the error case when going beyond that. Ensure that we can
handle values up to that point and generate errors beyond.

Reviewed by Tony Chang.

Test: storage/indexeddb/key-generator.html

* Modules/indexeddb/IDBBackingStore.h:
(IDBBackingStore):
* Modules/indexeddb/IDBLevelDBBackingStore.cpp:
(WebCore::IDBLevelDBBackingStore::nextAutoIncrementNumber):
* Modules/indexeddb/IDBLevelDBBackingStore.h:
(IDBLevelDBBackingStore):
* Modules/indexeddb/IDBObjectStoreBackendImpl.cpp:
(WebCore::IDBObjectStoreBackendImpl::putInternal):
(WebCore::IDBObjectStoreBackendImpl::genAutoIncrementKey):
* Modules/indexeddb/IDBObjectStoreBackendImpl.h:
(IDBObjectStoreBackendImpl):

LayoutTests:

Reviewed by Tony Chang.

* storage/indexeddb/key-generator-expected.txt:
* storage/indexeddb/resources/key-generator.js:
(get defineTest):


Canonical link: https://commits.webkit.org/102980@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115902 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/indexeddb/IDBLevelDBBackingStore.cpp
@@ -602,7 +602,7 @@ void IDBLevelDBBackingStore::deleteObjectStoreRecord(int64_t databaseId, int64_t
     m_currentTransaction->remove(existsEntryKey);
 }
 
-double IDBLevelDBBackingStore::nextAutoIncrementNumber(int64_t databaseId, int64_t objectStoreId)
+int64_t IDBLevelDBBackingStore::nextAutoIncrementNumber(int64_t databaseId, int64_t objectStoreId)
 {
     ASSERT(m_currentTransaction);
     const Vector<char> startKey = ObjectStoreDataKey::encode(databaseId, objectStoreId, minIDBKey());
@@ -610,9 +610,10 @@ double IDBLevelDBBackingStore::nextAutoIncrementNumber(int64_t databaseId, int64
 
     OwnPtr<LevelDBIterator> it = m_currentTransaction->createIterator();
 
-    int maxNumericKey = 0;
+    int64_t maxNumericKey = 0;
 
-    // FIXME: Be more efficient: seek to something after the object store data, then reverse.
+    // FIXME: This does a forward scan over all keys. Improve it.
+    // Since all dates > all numbers, create Date(-Infinity) and seek backwards.
 
     for (it->seek(startKey); it->isValid() && compareKeys(it->key(), stopKey) < 0; it->next()) {
         const char *p = it->key().begin();
